Add Go Protobuf example (#14)
As discussed in https://pantsbuild.slack.com/archives/C0D7TNJHL/p1650497897830709, a downside of Go is that it currently requires you to have go installed on your machine. That makes it harder for people to try out Python. So, by default, we do not run Go, but make it easy to turn on.